requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --resolver=backtracking
#
attrs==22.1.0
    # via
    #   cattrs
    #   jsii
aws-cdk-aws-lambda-python-alpha==2.50.0a0
    # via -r requirements.in
aws-cdk-lib==2.50.0
    # via
    #   -r requirements.in
    #   aws-cdk-aws-lambda-python-alpha
boto3==1.26.10
    # via -r requirements.in
boto3-stubs==1.26.10
    # via types-boto3
botocore==1.29.10
    # via
    #   boto3
    #   s3transfer
botocore-stubs==1.29.9
    # via
    #   boto3-stubs
    #   types-botocore
cattrs==22.2.0
    # via jsii
constructs==10.1.161
    # via
    #   -r requirements.in
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
exceptiongroup==1.0.4
    # via cattrs
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsii==1.71.0
    # via
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   constructs
numpy==1.23.4
    # via
    #   -r requirements.in
    #   pyarrow
publication==0.0.3
    # via
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   constructs
    #   jsii
pyarrow==10.0.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   jsii
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   constructs
    #   jsii
types-awscrt==0.15.3
    # via botocore-stubs
types-boto3==1.0.2
    # via -r requirements.in
types-botocore==1.0.2
    # via -r requirements.in
types-s3transfer==0.6.0.post4
    # via boto3-stubs
typing-extensions==4.4.0
    # via
    #   boto3-stubs
    #   jsii
urllib3==1.26.12
    # via botocore
